Fix "undefined authentication.QRConfig" error. (#152)
After the refactor to use fact requests exclusively, all QR authentication requests now require a fact.QRConfig object to be set up. This adds an extra dependency for developers using this functionality.

This pull request creates an alias of fact.QRConfig within the authentication package, making it easier to use for developers and resolving the "undefined authentication.QRConfig" error.